api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
maaker-excalidraw-mcp>=0.5.1
